+++ test/Util/TestUtilString.cc
@@ -1,6 +1,12 @@
+/* This file is part of the CARTA Image Viewer: https://github.com/CARTAvis/carta-backend
+   Copyright 2018- Academia Sinica Institute of Astronomy and Astrophysics (ASIAA),
+   Associated Universities, Inc. (AUI) and the Inter-University Institute for Data Intensive Astronomy (IDIA)
+   SPDX-License-Identifier: GPL-3.0-or-later
+*/
+
 #include <gtest/gtest.h>
-#include <vector>
 #include <string>
+#include <vector>
 #include "Util/String.h"
 
 TEST(StringUtilTest, SplitString_Basic) {
@@ -52,7 +58,6 @@ TEST(StringUtilTest, ConstantTimeStringCompare_DifferentStrings) {
 
 TEST(StringUtilTest, SafeStringEscape_EscapesSpecialCharacters) {
     EXPECT_EQ(SafeStringEscape("hello world!"), "hello%20world%21");
-    // EXPECT_EQ(SafeStringEscape("email@example.com"), "email@example.com");
 }
 
 TEST(StringUtilTest, SafeStringUnescape_UnescapesProperly) {
@@ -74,4 +79,32 @@ TEST(StringUtilTest, StringToInt_InvalidInput) {
 TEST(StringUtilTest, StringToInt_OutOfRange) {
     int value;
     EXPECT_FALSE(StringToInt("999999999999999999999", value));
+}
+
+TEST(StringUtilTest, StringCompare) {
+    EXPECT_TRUE(ConstantTimeStringCompare("hello world", "hello world"));
+    EXPECT_FALSE(ConstantTimeStringCompare("hello w1rld", "hello world"));
+    EXPECT_FALSE(ConstantTimeStringCompare("hello w1rld", "hello w2rld"));
+    EXPECT_FALSE(ConstantTimeStringCompare("hello w", "hello world"));
+    EXPECT_TRUE(ConstantTimeStringCompare("", ""));
+    EXPECT_FALSE(ConstantTimeStringCompare("hello world", ""));
+    EXPECT_FALSE(ConstantTimeStringCompare("", "hello world"));
+}
+
+TEST(StringUtilTest, HasSuffixCaseSensitive) {
+    EXPECT_TRUE(HasSuffix("test.fits", ".fits", true));
+    EXPECT_FALSE(HasSuffix("test.FITS", ".fits", true));
+    EXPECT_FALSE(HasSuffix("test.fits", ".FITS", true));
+    EXPECT_FALSE(HasSuffix("test.fits", ".xml", true));
+    EXPECT_TRUE(HasSuffix("test.fits.gz", ".fits.gz", true));
+    EXPECT_FALSE(HasSuffix("test.fits.gz", ".fits", true));
+}
+
+TEST(StringUtilTest, HasSuffixCaseInsensitive) {
+    EXPECT_TRUE(HasSuffix("test.fits", ".fits"));
+    EXPECT_TRUE(HasSuffix("test.FITS", ".fits"));
+    EXPECT_TRUE(HasSuffix("test.fits", ".FITS"));
+    EXPECT_FALSE(HasSuffix("test.fits", ".xml"));
+    EXPECT_TRUE(HasSuffix("test.fits.gz", ".fits.gz"));
+    EXPECT_FALSE(HasSuffix("test.fits.gz", ".fits"));
 }
